% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplesubstitution.R
\name{list_parameters}
\alias{list_parameters}
\title{Show parameters in a string}
\usage{
list_parameters(s, pattern = pp)
}
\arguments{
\item{s}{a string with some parameters.}
}
\value{
A character vector of parameters.
}
\description{
Show parameters in a string
}
\details{
By default, params are enclosed like so: `{$param}`.
This works fine for SQL. If you _need_ to change it, you can modify the pattern argument.
}
\examples{
list_parameters("SELECT {column} from {table}")
}
