% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplesubstitution.R
\name{substitution}
\alias{substitution}
\title{Substitution}
\usage{
substitution(s, params, pattern = pp)
}
\arguments{
\item{s}{A string in which to make replacements}

\item{params}{A named list of substitution parameters}

\item{pattern}{The pattern that encodes parameters.}
}
\value{
A string with replacements made.
}
\description{
Substitution
}
\details{
Given a string with parameters encoded inside braces, \code{substitution} replaces the parameters with values given in the
`params` argument. This is mostly useful for parameterizing SQL code.
}
\examples{
substitution("SELECT {thing} FROM {place}", list(thing = "TOOLS", place = "TOOLSHED"))
}
